format code with extensions files
